remove NODE_ENV from generateGlobalJs and move it to esbuild's define field
