+++ templates/_worker.js/index.ts
@@ -2,6 +2,8 @@ import { handleRequest } from './handleRequest';
 import { adjustRequestForVercel } from './utils';
 import type { AsyncLocalStorage } from 'node:async_hooks';
 
+declare const NODE_ENV: string;
+
 declare const __CONFIG__: ProcessedVercelConfig;
 
 declare const __BUILD_OUTPUT__: VercelBuildOutput;
@@ -12,7 +14,7 @@ declare const __ENV_ALS__: AsyncLocalStorage<unknown> & {
 
 export default {
 	async fetch(request, env, ctx) {
-		return __ENV_ALS__.run({ ...env, NODE_ENV: __ENV_ALS__.NODE_ENV }, () => {
+		return __ENV_ALS__.run({ ...env, NODE_ENV }, () => {
 			const adjustedRequest = adjustRequestForVercel(request);
 
 			return handleRequest(
